Update README.md
The changes are due to an extension of my mindset inorder to explore new changes. Optimise the resources i've ,for, i've them in limited. Create a strong Community that'll evolve our world, for an expansion of the mindset and attain the peak of acknowlegment. knowlegded our inner mind for him to be transformed to a supreme one , for knowledge and justice.